+++ hooks/lando-setup-build-engine-darwin.js
@@ -52,7 +52,7 @@ const getVersion = version => {
 /*
  * Helper to get docker compose v2 download url
  */
-const getEngineDownloadUrl = (id = '170107') => {
+const getEngineDownloadUrl = (id = '175267') => {
   const arch = process.arch === 'arm64' ? 'arm64' : 'amd64';
   return `https://desktop.docker.com/mac/main/${arch}/${id}/Docker.dmg`;
 };
@@ -95,10 +95,10 @@ module.exports = async (lando, options) => {
 
   // darwin install task
   options.tasks.push({
-    title: `Downloading build engine`,
+    title: 'Downloading build engine',
     id: 'setup-build-engine',
-    description: `@lando/build-engine (docker-desktop)`,
-    version: `docker-desktop ${install}`,
+    description: '@lando/build-engine (docker-desktop)',
+    version: `Docker Desktop ${install}`,
     hasRun: async () => {
       // start by looking at the engine install status
       // @NOTE: is this always defined?
@@ -106,10 +106,7 @@ module.exports = async (lando, options) => {
 
       // if we get here let's make sure the engine is on
       try {
-        await lando.engine.daemon.up();
-        const BuildEngine = require('../components/docker-engine');
-        const bengine = new BuildEngine(lando.config.buildEngine, {debug});
-        await bengine.info();
+        await lando.engine.daemon.up({max: 1, backoff: 1000});
         return true;
       } catch (error) {
         lando.log.debug('docker install task has not run %j', error);
@@ -166,7 +163,7 @@ module.exports = async (lando, options) => {
         result.download = ctx.download;
 
         // finish up
-        task.title = 'Installed build engine to /Applications/Docker.app';
+        task.title = 'Installed build engine (Docker Desktop) to /Applications/Docker.app';
         return result;
       } catch (error) {
         throw error;