+++ hooks/lando-setup-install-ca-win32.js
@@ -31,7 +31,11 @@ module.exports = async (lando, options) => {
       try {
         const fingerprint = require('../utils/get-fingerprint')(caCert);
         debug('computed sha1 fingerprint %o for ca %o', fingerprint, caCert);
-        return require('../utils/get-system-cas')().includes(fingerprint);
+
+        // get fingerprints
+        const winfps = await require('../utils/get-system-cas')();
+
+        return winfps.includes(fingerprint);
       } catch (error) {
         debug('error determining fingerprint of %o: %o %o', caCert, error.message, error.stack);
         return false;