+++ hooks/lando-setup-orchestrator.js
@@ -60,7 +60,7 @@ module.exports = async (lando, options) => {
     title: `Downloading orchestrator`,
     id: 'setup-orchestrator',
     description: '@lando/orchestrator (docker-compose)',
-    version: `docker-compose v${orchestrator}`,
+    version: `Docker Compose v${orchestrator}`,
     hasRun: async () => {
       return !!orchestratorBin && typeof orchestrator === 'string' && fs.existsSync(dest);
     },
@@ -74,7 +74,7 @@ module.exports = async (lando, options) => {
       const download = require('../utils/download-x')(url, {debug, dest, test: ['--version']});
       // success
       download.on('done', data => {
-        task.title = `Installed orchestrator to ${dest}`;
+        task.title = `Installed orchestrator (Docker Compose) to ${dest}`;
         resolve(data);
       });
       // handle errors