+++ plugins/networking/index.js
@@ -53,13 +53,21 @@ module.exports = lando => {
     // skip the installation of the network if set
     if (options.skipNetworking) return;
 
+    // we need access to dat socket for this to work
+    const dependsOn = ['linux', 'wsl']
+      .includes(lando.config.os.landoPlatform) ? ['setup-build-engine-group', 'setup-build-engine'] : ['setup-build-engine'];
+
     options.tasks.push({
       title: `Creating Landonet`,
       id: 'create-landonet',
-      dependsOn: ['setup-build-engine'],
+      dependsOn,
       description: '@lando/landonet',
       comments: {
-        'NOT INSTALLED': 'Will create Landonet',
+        'NOT INSTALLED': 'Will create LandoNet',
+      },
+      skip: () => {
+        if (!['linux', 'wsl'].includes(lando.config.os.landoPlatform)) return false;
+        return !require('../../utils/is-group-member')('docker');
       },
       hasRun: async () => {
         // if docker isnt even installed then this is easy
@@ -67,8 +75,8 @@ module.exports = lando => {
 
         // otherwise attempt to sus things out
         try {
+          await lando.engine.daemon.up({max: 3, backoff: 1000});
           const landonet = lando.engine.getNetwork(lando.config.networkBridge);
-          await lando.engine.daemon.up();
           await landonet.inspect();
           return lando.versions.networking > 1;
         } catch (error) {
@@ -82,7 +90,7 @@ module.exports = lando => {
         const daemon = new LandoDaemon(lando.cache, lando.events, undefined, lando.log);
 
         // we need docker up for this
-        await daemon.up();
+        await daemon.up({max: 5, backoff: 1000});
 
         // if we are v1 then disconnect and remove for upgrade
         if (lando.versions.networking === 1) {