+++ utils/get-win32-envvar-from-wsl.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = async (varname, {debug} = {}) => {
+  const args = ['-Command', `[Environment]::GetEnvironmentVariable('${varname}')`];
+  const {stdout} = await require('./run-command')('powershell.exe', args, {debug});
+  return stdout.trim();
+};