+++ utils/wslpath-2-winpath.js
@@ -1,0 +1,8 @@
+'use strict';
+
+module.exports = async (path, {
+  debug = require('debug')('@lando/wslpath-2-winpath'),
+} = {}) => {
+  const {stdout} = await require('./run-command')('wslpath', ['-w', path], {debug});
+  return stdout.trim();
+};